feat: refactor API key schemas into separate files and update controller imports
